+++ flash/core/model.py
@@ -40,7 +40,8 @@ from flash.core.data.auto_dataset import BaseAutoDataset
 from flash.core.data.data_pipeline import DataPipeline, DataPipelineState
 from flash.core.data.data_source import DataSource
 from flash.core.data.io.output import Output
-from flash.core.data.process import Deserializer, DeserializerMapping, Postprocess, Preprocess
+from flash.core.data.io.output_transform import OutputTransform
+from flash.core.data.process import Deserializer, DeserializerMapping, Preprocess
 from flash.core.data.properties import ProcessState
 from flash.core.optimizers.optimizers import _OPTIMIZERS_REGISTRY
 from flash.core.optimizers.schedulers import _SCHEDULERS_REGISTRY
@@ -57,8 +58,8 @@ from flash.core.utilities.types import (
     METRICS_TYPE,
     MODEL_TYPE,
     OPTIMIZER_TYPE,
+    OUTPUT_TRANSFORM_TYPE,
     OUTPUT_TYPE,
-    POSTPROCESS_TYPE,
     PREPROCESS_TYPE,
 )
 
@@ -318,7 +319,8 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
         deserializer: Either a single :class:`~flash.core.data.process.Deserializer` or a mapping of these to
             deserialize the input
         preprocess: :class:`~flash.core.data.process.Preprocess` to use as the default for this task.
-        postprocess: :class:`~flash.core.data.process.Postprocess` to use as the default for this task.
+        output_transform: :class:`~flash.core.data.io.output_transform.OutputTransform` to use as the default for this
+            task.
         output: The :class:`~flash.core.data.io.output.Output` to use when formatting prediction outputs.
     """
 
@@ -337,7 +339,7 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
         metrics: METRICS_TYPE = None,
         deserializer: DESERIALIZER_TYPE = None,
         preprocess: PREPROCESS_TYPE = None,
-        postprocess: POSTPROCESS_TYPE = None,
+        output_transform: OUTPUT_TRANSFORM_TYPE = None,
         output: OUTPUT_TYPE = None,
     ):
         super().__init__()
@@ -356,7 +358,7 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
 
         self._deserializer: Optional[Deserializer] = None
         self._preprocess: Optional[Preprocess] = preprocess
-        self._postprocess: Optional[Postprocess] = postprocess
+        self._output_transform: Optional[OutputTransform] = output_transform
         self._output: Optional[Output] = None
 
         # Explicitly set the output to call the setter
@@ -497,7 +499,7 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
         x = data_pipeline.device_preprocessor(running_stage)(x)
         x = x[0] if isinstance(x, list) else x
         predictions = self.predict_step(x, 0)  # batch_idx is always 0 when running with `model.predict`
-        predictions = data_pipeline.postprocessor(running_stage)(predictions)
+        predictions = data_pipeline.output_transform_processor(running_stage)(predictions)
         return predictions
 
     def predict_step(self, batch: Any, batch_idx: int, dataloader_idx: int = 0) -> Any:
@@ -568,29 +570,31 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
     def _resolve(
         old_deserializer: Optional[Deserializer],
         old_preprocess: Optional[Preprocess],
-        old_postprocess: Optional[Postprocess],
+        old_output_transform: Optional[OutputTransform],
         old_output: Optional[Output],
         new_deserializer: Optional[Deserializer],
         new_preprocess: Optional[Preprocess],
-        new_postprocess: Optional[Postprocess],
+        new_output_transform: Optional[OutputTransform],
         new_output: Optional[Output],
-    ) -> Tuple[Optional[Deserializer], Optional[Preprocess], Optional[Postprocess], Optional[Output]]:
-        """Resolves the correct :class:`~flash.core.data.process.Preprocess`, :class:`~flash.core.data.process.Postprocess`, and
-        :class:`~flash.core.data.io.output.Output` to use, choosing ``new_*`` if it is not None or a base class
-        (:class:`~flash.core.data.process.Preprocess`, :class:`~flash.core.data.process.Postprocess`, or
-        :class:`~flash.core.data.io.output.Output`) and ``old_*`` otherwise.
+    ) -> Tuple[Optional[Deserializer], Optional[Preprocess], Optional[OutputTransform], Optional[Output]]:
+        """Resolves the correct :class:`~flash.core.data.process.Preprocess`,
+        :class:`~flash.core.data.io.output_transform.OutputTransform`, and :class:`~flash.core.data.io.output.Output` to
+        use, choosing ``new_*`` if it is not None or a base class (:class:`~flash.core.data.process.Preprocess`,
+        :class:`~flash.core.data.io.output_transform.OutputTransform`, or :class:`~flash.core.data.io.output.Output`)
+        and ``old_*`` otherwise.
 
         Args:
             old_preprocess: :class:`~flash.core.data.process.Preprocess` to be overridden.
-            old_postprocess: :class:`~flash.core.data.process.Postprocess` to be overridden.
+            old_output_transform: :class:`~flash.core.data.io.output_transform.OutputTransform` to be overridden.
             old_output: :class:`~flash.core.data.io.output.Output` to be overridden.
             new_preprocess: :class:`~flash.core.data.process.Preprocess` to override with.
-            new_postprocess: :class:`~flash.core.data.process.Postprocess` to override with.
+            new_output_transform: :class:`~flash.core.data.io.output_transform.OutputTransform` to override with.
             new_output: :class:`~flash.core.data.io.output.Output` to override with.
 
         Returns:
-            The resolved :class:`~flash.core.data.process.Preprocess`, :class:`~flash.core.data.process.Postprocess`,
-            and :class:`~flash.core.data.io.output.Output`.
+            The resolved :class:`~flash.core.data.process.Preprocess`,
+            :class:`~flash.core.data.io.output_transform.OutputTransform`, and
+            :class:`~flash.core.data.io.output.Output`.
         """
         deserializer = old_deserializer
         if new_deserializer is not None and type(new_deserializer) != Deserializer:
@@ -600,15 +604,15 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
         if new_preprocess is not None and type(new_preprocess) != Preprocess:
             preprocess = new_preprocess
 
-        postprocess = old_postprocess
-        if new_postprocess is not None and type(new_postprocess) != Postprocess:
-            postprocess = new_postprocess
+        output_transform = old_output_transform
+        if new_output_transform is not None and type(new_output_transform) != OutputTransform:
+            output_transform = new_output_transform
 
         output = old_output
         if new_output is not None and type(new_output) != Output:
             output = new_output
 
-        return deserializer, preprocess, postprocess, output
+        return deserializer, preprocess, output_transform, output
 
     @torch.jit.unused
     @property
@@ -669,7 +673,7 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
         data_pipeline: Optional[DataPipeline] = None,
     ) -> Optional[DataPipeline]:
         """Build a :class:`.DataPipeline` incorporating available
-        :class:`~flash.core.data.process.Preprocess` and :class:`~flash.core.data.process.Postprocess`
+        :class:`~flash.core.data.process.Preprocess` and :class:`~flash.core.data.io.output_transform.OutputTransform`
         objects. These will be overridden in the following resolution order (lowest priority first):
 
         - Lightning ``Datamodule``, either attached to the :class:`.Trainer` or to the :class:`.Task`.
@@ -682,12 +686,13 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
                 the current data source format used.
             deserializer: deserializer to use
             data_pipeline: Optional highest priority source of
-                :class:`~flash.core.data.process.Preprocess` and :class:`~flash.core.data.process.Postprocess`.
+                :class:`~flash.core.data.process.Preprocess` and
+                :class:`~flash.core.data.io.output_transform.OutputTransform`.
 
         Returns:
             The fully resolved :class:`.DataPipeline`.
         """
-        deserializer, old_data_source, preprocess, postprocess, output = None, None, None, None, None
+        deserializer, old_data_source, preprocess, output_transform, output = None, None, None, None, None
 
         # Datamodule
         datamodule = None
@@ -699,32 +704,32 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
         if getattr(datamodule, "data_pipeline", None) is not None:
             old_data_source = getattr(datamodule.data_pipeline, "data_source", None)
             preprocess = getattr(datamodule.data_pipeline, "_preprocess_pipeline", None)
-            postprocess = getattr(datamodule.data_pipeline, "_postprocess_pipeline", None)
+            output_transform = getattr(datamodule.data_pipeline, "_output_transform", None)
             output = getattr(datamodule.data_pipeline, "_output", None)
             deserializer = getattr(datamodule.data_pipeline, "_deserializer", None)
 
         # Defaults / task attributes
-        deserializer, preprocess, postprocess, output = Task._resolve(
+        deserializer, preprocess, output_transform, output = Task._resolve(
             deserializer,
             preprocess,
-            postprocess,
+            output_transform,
             output,
             self._deserializer,
             self._preprocess,
-            self._postprocess,
+            self._output_transform,
             self._output,
         )
 
         # Datapipeline
         if data_pipeline is not None:
-            deserializer, preprocess, postprocess, output = Task._resolve(
+            deserializer, preprocess, output_transform, output = Task._resolve(
                 deserializer,
                 preprocess,
-                postprocess,
+                output_transform,
                 output,
                 getattr(data_pipeline, "_deserializer", None),
                 getattr(data_pipeline, "_preprocess_pipeline", None),
-                getattr(data_pipeline, "_postprocess_pipeline", None),
+                getattr(data_pipeline, "_output_transform", None),
                 getattr(data_pipeline, "_output", None),
             )
 
@@ -739,7 +744,7 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
         if deserializer is None or type(deserializer) is Deserializer:
             deserializer = getattr(preprocess, "deserializer", deserializer)
 
-        data_pipeline = DataPipeline(data_source, preprocess, postprocess, deserializer, output)
+        data_pipeline = DataPipeline(data_source, preprocess, output_transform, deserializer, output)
         self._data_pipeline_state = self._data_pipeline_state or DataPipelineState()
         self.attach_data_pipeline_state(self._data_pipeline_state)
         self._data_pipeline_state = data_pipeline.initialize(self._data_pipeline_state)
@@ -763,14 +768,14 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
     @torch.jit.unused
     @data_pipeline.setter
     def data_pipeline(self, data_pipeline: Optional[DataPipeline]) -> None:
-        self._deserializer, self._preprocess, self._postprocess, self.output = Task._resolve(
+        self._deserializer, self._preprocess, self._output_transform, self.output = Task._resolve(
             self._deserializer,
             self._preprocess,
-            self._postprocess,
+            self._output_transform,
             self._output,
             getattr(data_pipeline, "_deserializer", None),
             getattr(data_pipeline, "_preprocess_pipeline", None),
-            getattr(data_pipeline, "_postprocess_pipeline", None),
+            getattr(data_pipeline, "_output_transform", None),
             getattr(data_pipeline, "_output", None),
         )
 
@@ -785,8 +790,8 @@ class Task(DatasetProcessor, ModuleWrapperBase, LightningModule, metaclass=Check
 
     @torch.jit.unused
     @property
-    def postprocess(self) -> Postprocess:
-        return getattr(self.data_pipeline, "_postprocess_pipeline", None)
+    def output_transform(self) -> OutputTransform:
+        return getattr(self.data_pipeline, "_output_transform", None)
 
     def on_train_dataloader(self) -> None:
         if self.data_pipeline is not None: